# Assignments/Assignment 8/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "User Management API",
      "description": "This API allows user registration, authentication, and image upload.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:5000",
        "description": "Local server"
      }
    ],
    "paths": {
      "/user/create": {
        "post": {
          "summary": "Create a new user",
          "tags": ["User"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fullName": { "type": "string", "example": "John Doe" },
                    "email": { "type": "string", "example": "john@example.com" },
                    "password": { "type": "string", "example": "Secure@123" }
                  }
                }
              }
            }
          },
          "responses": {
            "201": { "description": "User created successfully" },
            "400": { "description": "Validation failed" }
          }
        }
      },
      "/user/login": {
        "post": {
          "summary": "Login user",
          "tags": ["User"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string", "example": "john@example.com" },
                    "password": { "type": "string", "example": "Secure@123" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Login successful" },
            "401": { "description": "Invalid credentials" }
          }
        }
      },
      "/user/getAll": {
        "get": {
          "summary": "Retrieve all users",
          "tags": ["User"],
          "responses": {
            "200": { "description": "List of all users" }
          }
        }
      },
      "/user/uploadImage": {
        "post": {
          "summary": "Upload user image",
          "tags": ["User"],
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string", "example": "john@example.com" },
                    "image": { "type": "string", "format": "binary" }
                  }
                }
              }
            }
          },
          "responses": {
            "201": { "description": "Image uploaded successfully" },
            "400": { "description": "Invalid file format" },
            "404": { "description": "User not found" }
          }
        }
      }
    }
  }
  